+++ src/app/shared/model/dialogflowForwarderRequest.ts
@@ -1,0 +1,3 @@
+export class DialogflowForwarderRequest {
+  constructor(public query: string, public sessionId: string) {}
+}