+++ src/app/shared/service/chat/chat.service.ts
@@ -2,6 +2,7 @@ import { HttpClient } from "@angular/common/http";
 import { Injectable } from "@angular/core";
 import { ReplaySubject } from "rxjs";
 import { v4 as uuid } from "uuid";
+import { DialogflowForwarderRequest } from "../../model/dialogflowForwarderRequest";
 import { Message, MessageSender } from "../../model/message";
 import { dialogflowForwarderUrl } from "../../../../assets/secrets";
 
@@ -19,20 +20,38 @@ export class ChatService {
     this.conversation.next(message);
   }
 
-  invokeDialogflowForwarder(message: Message) {
-    const data = {
+  chatWithBot(message: Message) {
+    this.addToConversation(message);
+
+    const dialogflowForwarderRequest: DialogflowForwarderRequest = {
       query: message.content,
       sessionId: this.sessionId
     };
 
-    this.http.post(dialogflowForwarderUrl, data).subscribe(res => {
-      const fulfillmentResponse = res[0].queryResult.fulfillmentMessages[0];
-      if (fulfillmentResponse.text) {
-        this.addToConversation(
-          new Message(fulfillmentResponse.text.text[0], MessageSender.Bot)
-        );
-      } else if (fulfillmentResponse.payload) {
-      }
-    });
+    this.invokeDialogflowForwarder(dialogflowForwarderRequest);
   }
+
+  invokeDialogflowForwarder(
+    dialogflowForwarderRequest: DialogflowForwarderRequest
+  ) {
+    this.http
+      .post(dialogflowForwarderUrl, dialogflowForwarderRequest)
+      .subscribe(response => {
+        const fulfillmentMessage =
+          response[0].queryResult.fulfillmentMessages[0];
+        if (fulfillmentMessage.text) {
+          this.handleTextFulfillment(fulfillmentMessage);
+        } else if (fulfillmentMessage.payload) {
+          this.handlePayloadFulfillment(fulfillmentMessage);
+        }
+      });
+  }
+
+  handleTextFulfillment(fulfillmentMessage: any): void {
+    this.addToConversation(
+      new Message(fulfillmentMessage.text.text[0], MessageSender.Bot)
+    );
+  }
+
+  handlePayloadFulfillment(fulfillmentMessage: any): void {}
 }